Add external delegate (#194)
* Assistant checkpoint: Refactor external value methods into JoshExternalVisitor

Assistant generated file changes:
- src/main/java/org/joshsim/lang/interpret/visitor/delegates/JoshExternalVisitor.java: Create new JoshExternalVisitor delegate
- src/main/java/org/joshsim/lang/interpret/visitor/delegates/JoshValueVisitor.java: Remove external value methods from JoshValueVisitor
- src/main/java/org/joshsim/lang/interpret/visitor/JoshParserToMachineVisitor.java: Add JoshExternalVisitor import and field, Add JoshExternalVisitor field declaration, Initialize JoshExternalVisitor in constructor, Update external value method delegates
- src/test/java/org/joshsim/lang/interpret/visitor/delegates/JoshValueVisitorTest.java: Remove external value tests from JoshValueVisitorTest, Remove unused imports from JoshValueVisitorTest
- src/test/java/org/joshsim/lang/interpret/visitor/delegates/JoshExternalVisitorTest.java: Create test class for JoshExternalVisitor

---

User prompt:

@src/main/java/org/joshsim/lang/interpret/visitor/JoshParserToMachineVisitor.java uses @src/main/java/org/joshsim/lang/interpret/visitor/delegates/JoshValueVisitor.java through @src/main/java/org/joshsim/lang/interpret/visitor/delegates/DelegateToolbox.java. There are two methods in @src/main/java/org/joshsim/lang/interpret/visitor/delegates/JoshValueVisitor.java: visitExternalValue and visitExternalValueAtTime. Please pull those out into a new @src/main/java/org/joshsim/lang/interpret/visitor/delegates/JoshVisitorDelegate.java implementor called JoshExternalVisitor and please update visitExternalValue and visitExternalValueAtTime within @src/main/java/org/joshsim/lang/interpret/visitor/JoshParserToMachineVisitor.java. Please provide top of file JavaDoc and top of class JavaDoc like @src/main/java/org/joshsim/lang/interpret/visitor/delegates/JoshValueVisitor.java. Please also bring the JavaDoc for visitExternalValue and visitExternalValueAtTime over into JoshExternalVisitor. Finally, please update @src/test/java/org/joshsim/lang/interpret/visitor/delegates/JoshValueVisitorTest.java, moving tests for visitExternalValue and visitExternalValueAtTime into a new similarly structured test called JoshExternalVisitorTest.


* Minor style fix.

* Replit touch up.

* Fix imports on parser to machine visitor.

* Minor comment update.
