+++ replit.nix
@@ -1,4 +1,4 @@
-{pkgs}: {
+{ pkgs }: {
   deps = [
     pkgs.vim
     pkgs.unzip
