+++ src/test/java/org/joshsim/lang/interpret/visitor/delegates/JoshValueVisitorTest.java
@@ -14,8 +14,7 @@ import org.joshsim.engine.value.engine.EngineValueFactory;
 import org.joshsim.engine.value.type.EngineValue;
 import org.joshsim.lang.antlr.JoshLangParser.AllExpressionContext;
 import org.joshsim.lang.antlr.JoshLangParser.BoolContext;
-import org.joshsim.lang.antlr.JoshLangParser.ExternalValueAtTimeContext;
-import org.joshsim.lang.antlr.JoshLangParser.ExternalValueContext;
+
 import org.joshsim.lang.antlr.JoshLangParser.IdentifierContext;
 import org.joshsim.lang.antlr.JoshLangParser.NumberContext;
 import org.joshsim.lang.antlr.JoshLangParser.StringContext;
@@ -200,63 +199,5 @@ class JoshValueVisitorTest {
     verify(mockMachine).push(allString);
   }
 
-  @Test
-  void testVisitExternalValue() {
-    // Mock
-    ExternalValueContext context = mock(ExternalValueContext.class);
-    IdentifierContext nameContext = mock(IdentifierContext.class);
-
-    // Mock the identifier() method to return nameContext
-    context.name = nameContext;
-    when(nameContext.getText()).thenReturn("externalVar");
-
-    // Test
-    Fragment result = visitor.visitExternalValue(context);
-
-    // Validate
-    assertNotNull(result);
-    assertTrue(result instanceof ActionFragment);
-
-    EventHandlerAction action = result.getCurrentAction();
-    assertNotNull(action);
-
-    EventHandlerMachine mockMachine = mock(EventHandlerMachine.class);
-    when(mockMachine.getStepCount()).thenReturn(42L);
-    org.mockito.Mockito.doNothing().when(mockMachine).pushExternal("externalVar", 42L);
-
-    action.apply(mockMachine);
-
-    verify(mockMachine).getStepCount();
-    verify(mockMachine).pushExternal("externalVar", 42L);
-  }
-
-  @Test
-  void testVisitExternalValueAtTime() {
-    // Mock
-    ExternalValueAtTimeContext context = mock(ExternalValueAtTimeContext.class);
-    IdentifierContext nameContext = mock(IdentifierContext.class);
-    Token stepNode = mock(Token.class);
-
-    context.name = nameContext;
-    when(nameContext.getText()).thenReturn("externalVar");
-    context.step = stepNode;
-    when(stepNode.getText()).thenReturn("123");
-
-    // Test
-    Fragment result = visitor.visitExternalValueAtTime(context);
-
-    // Validate
-    assertNotNull(result);
-    assertTrue(result instanceof ActionFragment);
-
-    EventHandlerAction action = result.getCurrentAction();
-    assertNotNull(action);
-
-    EventHandlerMachine mockMachine = mock(EventHandlerMachine.class);
-    org.mockito.Mockito.doNothing().when(mockMachine).pushExternal("externalVar", 123L);
-
-    action.apply(mockMachine);
-
-    verify(mockMachine).pushExternal("externalVar", 123L);
-  }
+  
 }
